Move image->rune conversion helpers to internal
